Update functions using `update_function`
When updating a function, we drop it first. This will cause an error if the
function did not previously exist, which is what we want.

In addition to being safer than blindly `CREATE OR UPDATE`ing, this also
works across every relational database I know of that supports functions,
where `CREATE OR UPDATE` does not.