Order triggers and functions by Postgres oid
Our previous strategy of relying on postgres to give us the
triggers/functions in the correct order was incorrect.
Those tables have no specific ordering so the results can come back in
any order.

Ordering by 'oid' is the simplest way to order everything by insertion
order. Dropping a view that depends on a child view also requires
dropping that child view, so ordering in this manner should be
sufficient.