Use db connection provided by Rails
Using the Rails db connection ensures we use valid database connection
information when checking the postgres server version.

Without specifying credentials, `PG` attempts to connect to the default
database (i.e., the default one for the current user) which, if it does
not exist, causes an error:

```
  Failure/Error: PG.connect.server_version >= 10_00_00

  PG::ConnectionBad:
    FATAL:  database "xxxxx" does not exist
```

By accessing the raw connection through `@connectable`, we avoid this
issue because we end up using the database configuration as specified by
the Rails app.