Fix use of ~> in Appraisal file
We were testing the wrong version of Rails because of a mistake with
the ~> operator. If we want to target Rails 5.0, we must say "~> 5.0.0"
not "~> 5.0". The ~> operator will choose the max possible value for the
last level specified, so "~> 5.0" is like saying "5.x". In effect we
were testing against Rails 5.2, then Rails 5.2, then Rails 5.2.