Support Ruby 3
The biggest change in Ruby 3 is [how keyword parameters are distinct
from positional parameters][1]. If we want to support 3 and pre-3, we
need to take a single hash as a positional parameter (`option = {}`). If
we use the `**option` syntax then things fail on the Ruby 3/Rails 6
combo because of how Rails calls our methods.

I also updated the Travis & Appraisal files. There is no need to support
Rails pre-6 on Ruby 3, since [Rails itself has no plans to support
it][2].

[1]: https://rubyreferences.github.io/rubychanges/3.0.html#keyword-arguments-are-now-fully-separated-from-positional-arguments
[2]: https://github.com/rails/rails/issues/40938